+++ packages/core/src/components/Table/Body/Cell/RowActionsCell/RowActionsCell.tsx
@@ -1,41 +1,55 @@
 import { ExpandMoreIcon } from '@medly-components/icons';
 import { WithStyle } from '@medly-components/utils';
-import { useCallback, memo } from 'react';
+import type { FC } from 'react';
+import { memo, useCallback } from 'react';
 import Checkbox from '../../../../Checkbox';
 import { LoadingDiv } from '../Styled';
 import { RowActionsCellStyled } from './RowActionsCell.styled';
 import { RowActionProps } from './types';
-import type { FC } from 'react';
 
 const Component: FC<RowActionProps> = memo(props => {
     const stopPropagation = useCallback((e: React.MouseEvent) => e.stopPropagation(), []);
+    const {
+        isLoading,
+        isRowExpanded,
+        isRowExpandable,
+        isRowSelectable,
+        isRowSelected,
+        isRowIndeterminate,
+        isRowSelectionDisabled,
+        onRowSelection,
+        onRowExpansionIconClick,
+        tableSize,
+        isGroupedTable,
+        showShadowAtRight,
+        isNavigated
+    } = props;
 
     return (
         <RowActionsCellStyled
             onClick={stopPropagation}
-            isLoading={props.isLoading}
-            isRowExpanded={props.isRowExpanded}
-            isRowExpandable={props.isRowExpandable}
-            isRowSelectable={props.isRowSelectable}
-            isRowSelected={props.isRowSelected}
-            tableSize={props.tableSize}
-            isGroupedTable={props.isGroupedTable}
-            showShadowAtRight={props.showShadowAtRight}
+            isLoading={isLoading}
+            isRowExpanded={isRowExpanded}
+            isRowExpandable={isRowExpandable}
+            isRowSelectable={isRowSelectable}
+            isRowSelected={isRowSelected}
+            tableSize={tableSize}
+            isGroupedTable={isGroupedTable}
+            showShadowAtRight={showShadowAtRight}
+            isNavigated={isNavigated}
         >
-            {props.isLoading ? (
+            {isLoading ? (
                 <LoadingDiv />
             ) : (
                 <>
-                    {props.isRowExpandable && (
-                        <ExpandMoreIcon size={props.tableSize === 'L' ? 'M' : 'S'} onClick={props.onRowExpansionIconClick} />
-                    )}
-                    {props.isRowSelectable && (
+                    {isRowExpandable && <ExpandMoreIcon size={tableSize === 'L' ? 'M' : 'S'} onClick={onRowExpansionIconClick} />}
+                    {isRowSelectable && (
                         <Checkbox
-                            size={props.tableSize === 'XS' ? props.tableSize : undefined}
-                            indeterminate={props.isRowIndeterminate}
-                            disabled={props.isRowSelectionDisabled}
-                            checked={props.isRowSelected}
-                            onChange={props.onRowSelection}
+                            size={tableSize === 'XS' ? tableSize : undefined}
+                            indeterminate={isRowIndeterminate}
+                            disabled={isRowSelectionDisabled}
+                            checked={isRowSelected}
+                            onChange={onRowSelection}
                             onClick={stopPropagation}
                         />
                     )}