+++ packages/layout/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.13](https://github.com/medly/medly-components/compare/@medly-components/layout@2.0.12...@medly-components/layout@2.0.13) (2021-12-05)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [2.0.12](https://github.com/medly/medly-components/compare/@medly-components/layout@2.0.11...@medly-components/layout@2.0.12) (2021-11-30)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [2.0.11](https://github.com/medly/medly-components/compare/@medly-components/layout@2.0.10...@medly-components/layout@2.0.11) (2021-11-30)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [2.0.10](https://github.com/medly/medly-components/compare/@medly-components/layout@2.0.9...@medly-components/layout@2.0.10) (2021-11-26)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [2.0.9](https://github.com/medly/medly-components/compare/@medly-components/layout@2.0.8...@medly-components/layout@2.0.9) (2021-11-26)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [2.0.8](https://github.com/medly/medly-components/compare/@medly-components/layout@2.0.7...@medly-components/layout@2.0.8) (2021-11-26)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
 ## [2.0.7](https://github.com/medly/medly-components/compare/@medly-components/layout@2.0.6...@medly-components/layout@2.0.7) (2021-11-25)
 
 **Note:** Version bump only for package @medly-components/layout