+++ packages/markdown/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.13](https://github.com/medly/medly-components/compare/@medly-components/markdown@0.4.12...@medly-components/markdown@0.4.13) (2021-11-30)
+
+**Note:** Version bump only for package @medly-components/markdown
+
+
+
+
+
 ## [0.4.12](https://github.com/medly/medly-components/compare/@medly-components/markdown@0.4.11...@medly-components/markdown@0.4.12) (2021-11-22)
 
 **Note:** Version bump only for package @medly-components/markdown