+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.1](https://github.com/medly/medly-components/compare/@medly-components/theme@3.0.0...@medly-components/theme@3.0.1) (2021-11-30)
+
+**Note:** Version bump only for package @medly-components/theme
+
+
+
+
+
 # [3.0.0](https://github.com/medly/medly-components/compare/@medly-components/theme@2.22.1...@medly-components/theme@3.0.0) (2021-11-18)